requirements.txt
django>=6.0,<6.1
djangorestframework>=3.17,<3.18
django-filter>=25.2,<25.3
drf-spectacular>=0.29,<0.30
mysqlclient>=2.2,<2.3
gunicorn>=25.3,<25.4
pytest>=9.0,<9.1
pytest-django>=4.12,<4.13
python-dotenv>=1.0,<2.0
mypy>=1.20,<2.0
django-stubs>=6.0,<7.0
djangorestframework-stubs>=3.16,<4.0
types-django-filter>=25.2,<26.0
black>=25.0,<26.0
